Transfer all admin functions from the Staff prototype to the Admin prototype
The functions unique to only admin were formerly placed on the staff prototype and this required an if statement that validates that the staff is an admin before the function can be performed. Now the functions unique to only admin were transferred to the admin prototype therefore rendering the if statements unnecessary for validation. The if statements were deleted. The test cases for the if statements were also deleted.